Update insertion sort, test sort button

+++ mvonbehr-hw1/mvonbehr-hw1/SortingAlgorithms.swift
@@ -9,7 +9,7 @@ import Foundation
 
 class SortingAlgorithms {
     
-    func insertionSort(array: [Int], onSwap: (([Int]) -> Void)) {
+    func insertionSort(array: [CGFloat], onSwap: (([CGFloat]) -> Void)) {
         var result = array
         
         for index in result.indices {
